R/atlantis_consumption.R {mfdbatlantis}	R Documentation

Atlantis stomach content parsers

Description

Routines to read and make use of the atlantis diet matrix

Usage

atlantis_stomach_matrix(adir)
atlantis_stomach_content(adir,
        consumption,
        predator_map,
        prey_map)

Arguments

adir
The atlantis directory created by atlantis_directory

consumption
Survey data generated by atlantis_fg_tracer with consumption = TRUE

predator_map
Vector of functional groups to MFDB species, e.g. c(FCD = 'COD')

prey_map
Vector of functional groups to MFDB species, e.g. c(FHE = 'HER'), any other species will be discarded

Value

atlantis_stomach_matrix returns the content of the Atlantis diet matrix, annotated with year/month.

atlantis_stomach_content expands the consumption data into a stomach survey, dividing the total consumed biomass into proportions based on the diet matrix.

[Package mfdbatlantis version 4.0-0 Index]