R/atlantis_directory.R {mfdbatlantis}	R Documentation

Open an Atlantis output directory

Description

Open Atlantis output directory, find required files

Usage

atlantis_directory(path,
        bgm_area = first_file(path, '*.bgm'),
        nc_catch = first_file(path, "*CATCH.nc"),
        nc_out = first_file(path, "*.nc"),
        nc_prod = first_file(path, "*PROD.nc"),
        txt_diet = first_file(path, "*DietCheck.txt"),
        xml_bio = first_file(path, '*[Bb]io*.xml'),
        xml_fg = first_file(path, '*Groups*.xml'),
        xml_fisheries = first_file(path, '*Fisheries*.xml'),
        xml_opt = first_file(path, '*[rR]un*.xml'),
        start_year = 1948)

Arguments

path
Directory containing model output

bgm_area
The BGM file that defines model boxes

nc_catch
The .nc file containing catch data

nc_out
The .nc file containing count and StructN

nc_prod
The .nc file containing Eat

txt_diet
The diet matrix

xml_bio
XML biology file

xml_fg
XML functional group definitions

xml_fisheries
XML fisheries definitions

xml_opt
XML run options

start_year
The year the model is assumed to start at

Details

All aguments apart from path are optional. If not supplied then it makes a best guess of the filename and errors if it can't be found.

Value

The path to the directory, annotated with filenames

[Package mfdbatlantis version 4.0-0 Index]