R/atlantis_fisheries.R {mfdbatlantis}	R Documentation

Atlantis stomach content parsers

Description

Routines to read and make use of the atlantis diet matrix

Usage

atlantis_fisheries(adir)
atlantis_fisheries_catch(adir,
        area_data,
        fishery)

Arguments

adir
The atlantis directory created by atlantis_directory

area_data
The results of atlantis_read_areas

fishery
A single fishery from atlantis_fisheries

Value

atlantis_fisheries returns a data.frame describing each fishery in the Atlantis output.

atlantis_fisheries_catch returns logbook output that can be ingested into MFDB.

[Package mfdbatlantis version 4.0-0 Index]