R/atlantis_tracer.R {mfdbatlantis}	R Documentation

Atlantis tracer parsers

Description

Routines to read and make use of the tracer output files

Usage

atlantis_temperature(adir,
        area_data)
atlantis_fg_tracer(adir,
        area_data,
        fg_group,
        consumption = FALSE)
atlantis_tracer_add_lengthgroups(
        tracer_data,
        length_group,
        sigma_per_cohort)
atlantis_tracer_survey_select(
        tracer_data,
        length_group,
        survey_suitability,
        survey_sigma)

Arguments

adir
The atlantis directory created by atlantis_directory

area_data
The area data returned by atlantis_read_areas

fg_group
One of the functional groups returned from atlantis_functional_groups

consumption
Include consumption data?

tracer_data
Output of atlantis_fg_tracer or atlantis_tracer_add_lengthgroups

length_group
Vector of c(min, min, min, ..., max) lengths

sigma_per_cohort
Vector of sigma per-cohort in age group

survey_suitability
Sutiability for vessel per-length-group

survey_sigma
Error rate per-length-group

Value

atlantis_temperature returns a data.frame of temperatures suitable for importing into MFDB.

atlantis_fg_tracer returns a data.frame containing total abundance/length/weight information, and optionally consumption/day for the given functional group. Can be imported into MFDB, but ideally a survey is simulated first.

atlantis_tracer_add_lengthgroups break up abundance from atlantis_fg_tracer into length groups, assuming a normal distribution around the Atlantis-reported mean.

atlantis_tracer_survey_select samples from atlantis_tracer_add_lengthgroups output based on the survey_suitability vector (possibly generated by rgadget), adding noise based on survey_sigma. The result is a data.frame that can be imported as a survey into MFDB.

[Package mfdbatlantis version 4.0-0 Index]